flaskDocker/flask/requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-JWT==0.3.2
Flask-JWT-Extended==3.21.0
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==3.0.1
marshmallow-sqlalchemy==0.17.0
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
toml==0.10.0
urllib3==1.25.3
uWSGI==2.0.18
Werkzeug==0.15.5
